Basic features are done!

+++ back.js
@@ -4,28 +4,33 @@ const fs = require('fs')
 const read = require('read')
 
 function ask(question) {
-  return read({ prompt: question, silent: false })
+  return read({ prompt: question, silent: true })
 }
 
 const cipherType = 'aes-256-cbc'
-const RESULT_CONTEXT_PATH = `./context.txt`
 const RESULT_ASSETS_PATH = `./assets.json`
-const ZIP_BACK_PATH = './back.zip'
+const ZIP_BACK_PATH = './result.zip'
 const FINAL_RESULT_PATH = './result'
 
 async function start() {
-  const context = fs.readFileSync(RESULT_CONTEXT_PATH, 'utf8')
-  const { enc: assetsEnc, iv: assetsIv } = JSON.parse(fs.readFileSync(RESULT_ASSETS_PATH, 'utf8'))
+  // Read origin enc and iv
+  const { enc: assetsEnc, iv: assetsIv, godWords } = JSON.parse(fs.readFileSync(RESULT_ASSETS_PATH, 'utf8'))
 
+  // Use password generate real enc and iv
   const password = await ask('Password: ')
-
   const { enc: backEnc, iv: backIv } = generateEncAndIv(password, { basicEncKey: assetsEnc, basicIv: assetsIv })
-  const keepIt = await welcomeBack(context, backEnc, backIv)
 
+  // Falling
+  const keepIt = await welcomeBack(godWords, backEnc, backIv).catch(() => null)
+  if (keepIt == null) return process.exit(2)
+
+  // Write them back, welcome back
   fs.writeFileSync(ZIP_BACK_PATH, keepIt, 'hex')
 
   fs.mkdirSync(FINAL_RESULT_PATH, { recursive: true })
   zipper.sync.unzip(ZIP_BACK_PATH).save(FINAL_RESULT_PATH)
+
+  console.log('\n\nWelcome back.\n')
 }
 start()
 
@@ -39,7 +44,7 @@ function welcomeBack(godWords, enc, iv) {
       const decipher = crypto.createDecipheriv(cipherType, enc, iv)
       return void resolve(decipher.update(godWords, 'hex', 'utf8') + decipher.final('utf8'))
     } catch (error) {
-      console.error('Decrypt failed')
+      console.error('\n\nYOU SHALL NOT PASS\n')
       return void reject(error)
     }
   })
